Rename AnnotatedInstanceSetHistory to ReportArchive
