+++ Google.Solutions.IapDesktop.Extensions.LogAnalysis.Test/Services/SchedulingReport/TestAnnotatedInstanceSetHistory.cs
@@ -21,13 +21,13 @@
 
 using Google.Solutions.Common.Locator;
 using Google.Solutions.IapDesktop.Extensions.LogAnalysis.History;
-using Google.Solutions.IapDesktop.Extensions.LogAnalysis.Services.UsageReport;
+using Google.Solutions.IapDesktop.Extensions.LogAnalysis.Services.SchedulingReport;
 using NUnit.Framework;
 using System;
 using System.IO;
 using System.Linq;
 
-namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.Test.Services.UsageReport
+namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.Test.Services.SchedulingReport
 {
     [TestFixture]
     public class TestAnnotatedInstanceSetHistory : FixtureBase
@@ -68,7 +68,7 @@ namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.Test.Services.Usage
                         })
                 });
 
-            var annotatedHistory = AnnotatedInstanceSetHistory.FromInstanceSetHistory(history);
+            var annotatedHistory = ReportArchive.FromInstanceSetHistory(history);
             annotatedHistory.AddLicenseAnnotation(
                 new ImageLocator("project-1", "windows"),
                 OperatingSystemTypes.Windows,
@@ -85,7 +85,7 @@ namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.Test.Services.Usage
 
                 memoryStream.Position = 0;
 
-                var restoredAnnotatedHistory = AnnotatedInstanceSetHistory
+                var restoredAnnotatedHistory = ReportArchive
                     .Deserialize(new StreamReader(memoryStream));
                 var restoredHistory = restoredAnnotatedHistory.History;
 
@@ -133,7 +133,7 @@ namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.Test.Services.Usage
 
             using (var reader = new StringReader(json))
             {
-                Assert.Throws<FormatException>(() => AnnotatedInstanceSetHistory.Deserialize(reader));
+                Assert.Throws<FormatException>(() => ReportArchive.Deserialize(reader));
             }
         }
 
@@ -149,7 +149,7 @@ namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.Test.Services.Usage
 
             using (var reader = new StringReader(json))
             {
-                Assert.Throws<FormatException>(() => AnnotatedInstanceSetHistory.Deserialize(reader));
+                Assert.Throws<FormatException>(() => ReportArchive.Deserialize(reader));
             }
         }
 
@@ -216,7 +216,7 @@ namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.Test.Services.Usage
 
             using (var reader = new StringReader(json))
             {
-                var restoredAnnotatedHistory = AnnotatedInstanceSetHistory.Deserialize(reader);
+                var restoredAnnotatedHistory = ReportArchive.Deserialize(reader);
                 var restoredHistory = restoredAnnotatedHistory.History;
 
                 Assert.AreEqual(new DateTime(2019, 12, 1, 0, 0, 0, DateTimeKind.Utc), restoredHistory.StartDate);
@@ -284,7 +284,7 @@ namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.Test.Services.Usage
                         })
                 });
 
-            var annotatedHistory = AnnotatedInstanceSetHistory.FromInstanceSetHistory(history);
+            var annotatedHistory = ReportArchive.FromInstanceSetHistory(history);
             Assert.IsTrue(annotatedHistory.GetInstances(
                     OperatingSystemTypes.Unknown, 
                     LicenseTypes.Unknown).Any());
@@ -322,7 +322,7 @@ namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.Test.Services.Usage
                         })
                 });
 
-            var annotatedHistory = AnnotatedInstanceSetHistory.FromInstanceSetHistory(history);
+            var annotatedHistory = ReportArchive.FromInstanceSetHistory(history);
             annotatedHistory.AddLicenseAnnotation(
                 new ImageLocator("project-1", "image"),
                 OperatingSystemTypes.Windows,
