+++ Google.Solutions.IapDesktop.Extensions.LogAnalysis/Services/SchedulingReport/ReportPaneViewModelBase.cs
@@ -25,7 +25,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.Services.UsageReport
+namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.Services.SchedulingReport
 {
 
     internal abstract class ReportPaneViewModelBase : ViewModelBase
